requirements.txt
asgiref==3.5.2
Django==3.2.9
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
